pyproject.toml
[project]
name = "receipts-api"
version = "0.1.0"
requires-python = ">=3.13.4"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.117.1",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.43",
]

[dependency-groups]
ci = [
    "pyright>=1.1.405",
    "ruff>=0.13.1",
]
dev = [
    "pre-commit>=4.3.0",
]
tests = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.uv]
required-version = "==0.7.19"

[tool.pyright]
include = ["."]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
extraPaths = ["."]
executionEnvironments = [
    { root = ".", extraPaths = ["."] }
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "E4", "E7", "E9", "W", "S", "I", "B", "SIM"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106", "S107"]

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"
